requirements.txt
appdirs
pywizlight
hue-api
websockets

